// libs/database/prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AREA_TYPE = {
  HECTAR: 'HECTAR',
  METER: 'METER',
  KM: 'KM'
} as const

export type AREA_TYPE = (typeof AREA_TYPE)[keyof typeof AREA_TYPE]